Extract nested class with other class nested to separate file. #1555
Fixes `ClassNestingDepth` inspection violation.

Description:
>Reports inner classes too deeply nested. Nesting inner classes inside inner classes is almost certain to be confusing, and is a good sign that refactoring may be necessary.
